+++ data/mamtainput.txt
@@ -1,4 +1,16 @@
 T|X|read book
-D| |return book|2019-10-15 1800
-E| |project meeting|2004-10-29 400|2-4pm
+D| |return book|OCTOBER 15 2019 1800
+E| |project meeting|OCTOBER 29 2004 400|2-4pm
 T|X|join sports club
+T| |borrow book 
+D| |return book|Sunday
+E| |project meeting|Mon 2pm|4pm
+T| |borrow book 
+D| |return book|Sunday
+E| |project meeting|Mon 2pm|4pm
+T| |borrow book 
+D| |return book|Sunday
+E| |project meeting|Mon 2pm|4pm
+T| |borrow book 
+D| |return book|Sunday
+E| |project meeting|Mon 2pm|4pm
